+++ packages/config/src/configs/outbox-runner-config.ts
@@ -1,0 +1,18 @@
+import { IsDefined, IsPositive, ValidateNested } from 'class-validator';
+import { LoggerConfig } from './logger-config.js';
+import { Type } from 'class-transformer';
+
+export class OutboxRunnerConfig {
+  @IsDefined()
+  @IsPositive()
+  batchIntervalMs!: number;
+
+  @IsDefined()
+  @IsPositive()
+  batchSize!: number;
+
+  @IsDefined()
+  @ValidateNested()
+  @Type(() => LoggerConfig)
+  logger!: LoggerConfig;
+}