+++ packages/contracts/src/user/user.command.ts
@@ -5,7 +5,7 @@
 // source: volontariapp/user/user.command.proto
 
 /* eslint-disable */
-import { OrganisationInfo } from "./user.js";
+import { OrganisationInfo } from './user.js';
 
 export interface SignUpCommand {
   email: string;
@@ -40,4 +40,38 @@ export interface LoginCommand {
 
 export interface RefreshTokenCommand {
   refreshToken: string;
+}
+
+export interface IncrementImpactScoreCommand {
+  userId: string;
+  scoreIncrement: number;
+}
+
+export interface CreateBadgeCommand {
+  name: string;
+  slug: string;
+  iconPath?: string | undefined;
+  description: string;
+}
+
+export interface DeleteBadgeCommand {
+  badgeId: string;
+}
+
+export interface UpdateBadgeCommand {
+  badgeId: string;
+  name?: string | undefined;
+  slug?: string | undefined;
+  iconPath?: string | undefined;
+  description?: string | undefined;
+}
+
+export interface AddBadgeToUserCommand {
+  userId: string;
+  badgeId: string;
+}
+
+export interface RemoveBadgeFromUserCommand {
+  userId: string;
+  badgeId: string;
 }