+++ packages/contracts/src/user/user.query.ts
@@ -5,12 +5,20 @@
 // source: volontariapp/user/user.query.proto
 
 /* eslint-disable */
-import { PaginationRequest } from "../common/pagination.js";
+import { PaginationRequest } from '../common/pagination.js';
 
 export interface GetUserQuery {
   userId: string;
 }
 
 export interface ListUsersQuery {
+  pagination: PaginationRequest | undefined;
+}
+
+export interface GetBadgeQuery {
+  badgeId: string;
+}
+
+export interface ListBadgesQuery {
   pagination: PaginationRequest | undefined;
 }